nvapi,nvofapi: Reorganize file layout
