+++ src/meson.build
@@ -2,17 +2,17 @@ nvapi_src = files([
   'util/util_string.cpp',
   'util/util_env.cpp',
   'util/util_log.cpp',
-  'sysinfo/vulkan.cpp',
-  'sysinfo/nvml.cpp',
-  'sysinfo/nvapi_output.cpp',
-  'sysinfo/nvapi_adapter.cpp',
-  'sysinfo/nvapi_adapter_registry.cpp',
+  'nvapi/lfx.cpp',
+  'nvapi/nvapi_d3d_instance.cpp',
+  'nvapi/nvapi_d3d_low_latency_device.cpp',
+  'nvapi/nvapi_d3d11_device.cpp',
+  'nvapi/nvapi_d3d12_device.cpp',
+  'nvapi/nvapi_adapter_registry.cpp',
+  'nvapi/nvapi_adapter.cpp',
+  'nvapi/nvapi_output.cpp',
+  'nvapi/vulkan.cpp',
+  'nvapi/nvml.cpp',
   'resource_factory.cpp',
-  'd3d/lfx.cpp',
-  'd3d/nvapi_d3d_instance.cpp',
-  'd3d/nvapi_d3d_low_latency_device.cpp',
-  'd3d11/nvapi_d3d11_device.cpp',
-  'd3d12/nvapi_d3d12_device.cpp',
   'nvapi_globals.cpp',
   'nvapi.cpp',
   'nvapi_d3d.cpp',
@@ -36,13 +36,13 @@ nvofapi_src = files([
   'util/util_string.cpp',
   'util/util_env.cpp',
   'util/util_log.cpp',
+  'nvofapi/nvofapi_image.cpp',
+  'nvofapi/nvofapi_common.cpp',
+  'nvofapi/nvofapi_d3d12_instance.cpp',
+  'nvofapi/nvofapi_vulkan_instance.cpp',
   'nvofapi.cpp',
   'nvofapi_d3d12.cpp',
   'nvofapi_vulkan.cpp',
-  'd3d12/nvofapi_d3d12_instance.cpp',
-  'vulkan/nvofapi_vulkan_instance.cpp',
-  'nvofapi/nvofapi_common.cpp',
-  'nvofapi/nvofapi_image.cpp',
 ])
 
 # Only build 64-bit versions of nvofapi
