+++ src/nvapi/nvapi_adapter.cpp
