+++ src/nvapi/nvapi_adapter.h
