+++ src/nvapi/nvapi_adapter_registry.cpp
@@ -1,7 +1,7 @@
 #include "nvapi_adapter_registry.h"
-#include "../d3d11/nvapi_d3d11_device.h"
-#include "../d3d12/nvapi_d3d12_device.h"
-#include "../d3d/nvapi_d3d_low_latency_device.h"
+#include "nvapi_d3d11_device.h"
+#include "nvapi_d3d12_device.h"
+#include "nvapi_d3d_low_latency_device.h"
 #include "../util/util_log.h"
 
 namespace dxvk {
