+++ src/nvapi/nvapi_adapter_registry.h
