+++ src/nvapi/nvapi_d3d11_device.h
