+++ src/nvapi/nvapi_d3d12_device.h
