+++ src/nvapi/nvapi_d3d_instance.h
