+++ src/nvapi/nvapi_d3d_low_latency_device.cpp
