+++ src/nvapi/nvapi_output.cpp
