+++ src/nvapi/nvapi_output.h
