+++ src/nvapi_d3d12.cpp
@@ -1,7 +1,7 @@
 #include "nvapi_private.h"
 #include "nvapi_globals.h"
-#include "d3d/nvapi_d3d_low_latency_device.h"
-#include "d3d12/nvapi_d3d12_device.h"
+#include "nvapi/nvapi_d3d_low_latency_device.h"
+#include "nvapi/nvapi_d3d12_device.h"
 #include "util/util_statuscode.h"
 #include "util/util_op_code.h"
 #include "util/util_pso_extension.h"
