+++ src/nvapi_globals.h
@@ -2,8 +2,8 @@
 
 #include "nvapi_private.h"
 #include "resource_factory.h"
-#include "sysinfo/nvapi_adapter_registry.h"
-#include "d3d/nvapi_d3d_instance.h"
+#include "nvapi/nvapi_adapter_registry.h"
+#include "nvapi/nvapi_d3d_instance.h"
 
 extern uint64_t initializationCount;
 extern std::unique_ptr<dxvk::ResourceFactory> resourceFactory;
