+++ tests/meson.build
@@ -2,17 +2,17 @@ nvapi_src = files([
   '../src/util/util_string.cpp',
   '../src/util/util_env.cpp',
   '../src/util/util_log.cpp',
-  '../src/sysinfo/vulkan.cpp',
-  '../src/sysinfo/nvml.cpp',
-  '../src/d3d/lfx.cpp',
-  '../src/d3d/nvapi_d3d_instance.cpp',
-  '../src/d3d/nvapi_d3d_low_latency_device.cpp',
-  '../src/sysinfo/nvapi_output.cpp',
-  '../src/sysinfo/nvapi_adapter.cpp',
-  '../src/sysinfo/nvapi_adapter_registry.cpp',
+  '../src/nvapi/vulkan.cpp',
+  '../src/nvapi/nvml.cpp',
+  '../src/nvapi/lfx.cpp',
+  '../src/nvapi/nvapi_d3d_instance.cpp',
+  '../src/nvapi/nvapi_d3d_low_latency_device.cpp',
+  '../src/nvapi/nvapi_output.cpp',
+  '../src/nvapi/nvapi_adapter.cpp',
+  '../src/nvapi/nvapi_adapter_registry.cpp',
   '../src/resource_factory.cpp',
-  '../src/d3d11/nvapi_d3d11_device.cpp',
-  '../src/d3d12/nvapi_d3d12_device.cpp',
+  '../src/nvapi/nvapi_d3d11_device.cpp',
+  '../src/nvapi/nvapi_d3d12_device.cpp',
   '../src/nvapi_globals.cpp',
   '../src/nvapi.cpp',
   '../src/nvapi_d3d.cpp',
