+++ tests/nvapi_d3d_mocks.h
@@ -2,7 +2,7 @@
 
 #include "nvapi_tests_private.h"
 #include "../src/interfaces/shared_interfaces.h"
-#include "../src/d3d/lfx.h"
+#include "../src/nvapi/lfx.h"
 
 class UnknownMock : public trompeloeil::mock_interface<IUnknown> {
     MAKE_MOCK2(QueryInterface, HRESULT(REFIID, void**), override);
