+++ tests/nvapi_sysinfo_mocks.h
@@ -2,8 +2,8 @@
 
 #include "nvapi_tests_private.h"
 #include "../src/interfaces/dxvk_interfaces.h"
-#include "../src/sysinfo/vulkan.h"
-#include "../src/sysinfo/nvml.h"
+#include "../src/nvapi/vulkan.h"
+#include "../src/nvapi/nvml.h"
 
 class IDXGIDxvkFactoryMock : public IDXGIFactory1, public IDXGIVkInteropFactory1 {};
